Update job status and implement bug fixes
The commit includes changes to reflect updated job statuses, bug fixes within the job service, and profile repositories. The update includes a verification step for jobId and checks for unpaid jobs. It also solves an existing bug detailed in the README file. Several changes were made to optimize the program, such as the validation of a job in the service before payment and improvements in transaction handling. Finally, the commit also downgrades the Sequelize version in the package for compatibility.